+++ project_composer/compose/__init__.py
@@ -2,6 +2,7 @@ from .base import ComposerBase
 from .classes import ClassComposer
 from .text import TextContentComposer
 from .django import ComposeDjangoSettings, ComposeDjangoUrls
+from .purge import PurgeApplications
 from .requirements import RequirementsComposer
 
 
@@ -11,5 +12,6 @@ __all__ = [
     "TextContentComposer",
     "ComposeDjangoSettings",
     "ComposeDjangoUrls",
+    "PurgeApplications",
     "RequirementsComposer",
 ]