+++ project_composer/defaults.py
@@ -1,0 +1,21 @@
+"""
+=================
+Default constants
+=================
+
+"""
+
+
+# Default manifest filepath
+DEFAULT_MANIFEST_PATH = "composition-manifest.json"
+
+
+# Default logger conf
+APP_LOGGER_CONF = (
+    "DEBUG",
+    "INFO",
+    "WARNING",
+    "ERROR",
+    "CRITICAL",
+    None
+)